requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
Pillow>=10.0.0
opencv-python>=4.8.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0